fix(fork-network): allow setting the genesis time in set-validators (#10966)
If this tools is to be used to prepare nodes for a mocknet (which we
want to do because it's much faster than starting from a giant records
json file), we need to set the genesis time to the same value on all
nodes. Otherwise they'll have slightly different genesis blocks and
refuse to connect to each other. So just add a flag to set-validators
that allows setting it, and keep the old behavior if it's not present.